# Fix #31, fix #28: correct verbalization of slowdown/speed up