Fixes 
- #31 
- #28 

Follows suggestion raised towards
- #29 (superseded)

New output (example):
```
  insert
    Set Bool: OK (1.80s)
      845  ns ±  63 ns, unchanged against baseline
    BoolSet:  OK (2.34s)
      138  ns ± 7.2 ns, faster than baseline by factor 6.13
```