@andreasabel 2022-05-22
LGTM, except for missing acknowledgement in commit, e.g. "jww Andreas Abel" or "with contributions by ..." etc.
Citations, acknowledgements and the like are the currency in science and volunteer work.
Next time...