@Bodigrim 2022-05-22
@andreasabel sorry, I appreciate our discussion in #28 a lot and it __is__ cited in the commit message. In terms of actual implementation fce531307fd58f2b94f81ce4c8445c693fae070b does not bear much resemblance to 003ad47 or 51ca6ce, and I did not read them beforehand.